changed type size_t to unsigned byte for loop

+++ includes/Eye.h
@@ -22,7 +22,7 @@ public:
     ~Eye();
 private:
     //Number of SERVOS used to handle the eye behaviour.
-    static const size_t NUM_OF_SERVOS_USED = 2;
+    static const uint8_t NUM_OF_SERVOS_USED = 2;
     // First idx is de horizontal servo controller.
     // Second idx is the vertical servo controller.
     Servo usedServos[NUM_OF_SERVOS_USED];
